% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrabble_score.R
\name{score}
\alias{score}
\title{Title}
\usage{
score(x, lang)
}
\arguments{
\item{x}{a character vector}

\item{lang}{The language to base the score on. Currently only German (de) supported}
}
\value{
The SCRABBLE score for the input character vector
}
\description{
Title
}
