% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supported_languages.R
\name{supported_languages}
\alias{supported_languages}
\title{Print supported languages}
\usage{
supported_languages()
}
\value{
a vector with all supported languages (abbreviations)
}
\description{
Print supported languages
}
