% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/which_languages.R
\name{which_languages}
\alias{which_languages}
\title{Print supported languages}
\usage{
which_languages()
}
\value{
a vector with all supported languages (abbreviations)
}
\description{
Print supported languages
}
